+++ spec/graphql/resolvers/integration_collection_mapping_resolver_spec.rb
@@ -2,15 +2,14 @@
 
 require 'rails_helper'
 
-RSpec.describe Resolvers::IntegrationMappings::MappingResolver, type: :graphql do
+RSpec.describe Resolvers::IntegrationCollectionMappingResolver, type: :graphql do
   let(:required_permission) { 'organization:integrations:view' }
   let(:query) do
     <<~GQL
-      query($integrationMappingId: ID!) {
-        integrationMapping(id: $integrationMappingId) {
+      query($collectionMappingId: ID!) {
+        integrationCollectionMapping(id: $collectionMappingId) {
           id
-          mappableId
-          mappableType
+          mappingType
           externalId
           externalAccountCode
           externalName
@@ -22,35 +21,34 @@ RSpec.describe Resolvers::IntegrationMappings::MappingResolver, type: :graphql d
   let(:membership) { create(:membership) }
   let(:organization) { membership.organization }
   let(:integration) { create(:netsuite_integration, organization:) }
-  let(:netsuite_mapping) { create(:netsuite_mapping, integration:) }
+  let(:netsuite_collection_mapping) { create(:netsuite_collection_mapping, integration:) }
 
   before do
-    netsuite_mapping
+    netsuite_collection_mapping
   end
 
   it_behaves_like 'requires current user'
   it_behaves_like 'requires current organization'
   it_behaves_like 'requires permission', 'organization:integrations:view'
 
-  it 'returns a single integration mapping' do
+  it 'returns a single integration collection mapping' do
     result = execute_graphql(
       current_user: membership.user,
       current_organization: organization,
       permissions: required_permission,
       query:,
-      variables: { integrationMappingId: netsuite_mapping.id },
+      variables: { collectionMappingId: netsuite_collection_mapping.id },
     )
 
-    integration_mapping_response = result['data']['integrationMapping']
+    integration_mapping_response = result['data']['integrationCollectionMapping']
 
     aggregate_failures do
-      expect(integration_mapping_response['id']).to eq(netsuite_mapping.id)
-      expect(integration_mapping_response['mappableId']).to eq(netsuite_mapping.mappable_id)
-      expect(integration_mapping_response['mappableType']).to eq(netsuite_mapping.mappable_type)
-      expect(integration_mapping_response['externalId']).to eq(netsuite_mapping.external_id)
-      expect(integration_mapping_response['externalName']).to eq(netsuite_mapping.external_name)
+      expect(integration_mapping_response['id']).to eq(netsuite_collection_mapping.id)
+      expect(integration_mapping_response['mappingType']).to eq(netsuite_collection_mapping.mapping_type)
+      expect(integration_mapping_response['externalId']).to eq(netsuite_collection_mapping.external_id)
+      expect(integration_mapping_response['externalName']).to eq(netsuite_collection_mapping.external_name)
       expect(integration_mapping_response['externalAccountCode'])
-        .to eq(netsuite_mapping.external_account_code)
+        .to eq(netsuite_collection_mapping.external_account_code)
     end
   end
 
@@ -61,7 +59,7 @@ RSpec.describe Resolvers::IntegrationMappings::MappingResolver, type: :graphql d
         current_organization: organization,
         permissions: required_permission,
         query:,
-        variables: { integrationMappingId: '123456' },
+        variables: { collectionMappingId: '123456' },
       )
 
       expect_graphql_error(
