+++ spec/graphql/resolvers/integration_mapping_resolver_spec.rb
@@ -2,14 +2,15 @@
 
 require 'rails_helper'
 
-RSpec.describe Resolvers::IntegrationCollectionMappings::CollectionMappingResolver, type: :graphql do
+RSpec.describe Resolvers::IntegrationMappingResolver, type: :graphql do
   let(:required_permission) { 'organization:integrations:view' }
   let(:query) do
     <<~GQL
-      query($collectionMappingId: ID!) {
-        collectionMapping(id: $collectionMappingId) {
+      query($integrationMappingId: ID!) {
+        integrationMapping(id: $integrationMappingId) {
           id
-          mappingType
+          mappableId
+          mappableType
           externalId
           externalAccountCode
           externalName
@@ -21,34 +22,35 @@ RSpec.describe Resolvers::IntegrationCollectionMappings::CollectionMappingResolv
   let(:membership) { create(:membership) }
   let(:organization) { membership.organization }
   let(:integration) { create(:netsuite_integration, organization:) }
-  let(:netsuite_collection_mapping) { create(:netsuite_collection_mapping, integration:) }
+  let(:netsuite_mapping) { create(:netsuite_mapping, integration:) }
 
   before do
-    netsuite_collection_mapping
+    netsuite_mapping
   end
 
   it_behaves_like 'requires current user'
   it_behaves_like 'requires current organization'
   it_behaves_like 'requires permission', 'organization:integrations:view'
 
-  it 'returns a single integration collection mapping' do
+  it 'returns a single integration mapping' do
     result = execute_graphql(
       current_user: membership.user,
       current_organization: organization,
       permissions: required_permission,
       query:,
-      variables: { collectionMappingId: netsuite_collection_mapping.id },
+      variables: { integrationMappingId: netsuite_mapping.id },
     )
 
-    integration_mapping_response = result['data']['collectionMapping']
+    integration_mapping_response = result['data']['integrationMapping']
 
     aggregate_failures do
-      expect(integration_mapping_response['id']).to eq(netsuite_collection_mapping.id)
-      expect(integration_mapping_response['mappingType']).to eq(netsuite_collection_mapping.mapping_type)
-      expect(integration_mapping_response['externalId']).to eq(netsuite_collection_mapping.external_id)
-      expect(integration_mapping_response['externalName']).to eq(netsuite_collection_mapping.external_name)
+      expect(integration_mapping_response['id']).to eq(netsuite_mapping.id)
+      expect(integration_mapping_response['mappableId']).to eq(netsuite_mapping.mappable_id)
+      expect(integration_mapping_response['mappableType']).to eq(netsuite_mapping.mappable_type)
+      expect(integration_mapping_response['externalId']).to eq(netsuite_mapping.external_id)
+      expect(integration_mapping_response['externalName']).to eq(netsuite_mapping.external_name)
       expect(integration_mapping_response['externalAccountCode'])
-        .to eq(netsuite_collection_mapping.external_account_code)
+        .to eq(netsuite_mapping.external_account_code)
     end
   end
 
@@ -59,7 +61,7 @@ RSpec.describe Resolvers::IntegrationCollectionMappings::CollectionMappingResolv
         current_organization: organization,
         permissions: required_permission,
         query:,
-        variables: { collectionMappingId: '123456' },
+        variables: { integrationMappingId: '123456' },
       )
 
       expect_graphql_error(
