+++ spec/graphql/resolvers/integration_mappings_resolver_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe Resolvers::IntegrationMappings::MappingsResolver, type: :graphql do
+RSpec.describe Resolvers::IntegrationMappingsResolver, type: :graphql do
   let(:required_permission) { 'organization:integrations:view' }
   let(:query) do
     <<~GQL
